+++ AdventOfCode/IntCode/Interfaces/IInputModule.cs
@@ -2,6 +2,6 @@ namespace AdventOfCode.IntCode.Interfaces
 {
     public interface IInputModule
     {
-        int InputCallback();
+        long InputCallback();
     }
 }
