+++ AdventOfCode/IntCode/Interfaces/IOutputModule.cs
@@ -2,6 +2,6 @@ namespace AdventOfCode.IntCode.Interfaces
 {
     public interface IOutputModule
     {
-        void OutputCallback(int val);
+        void OutputCallback(long val);
     }
 }
