+++ AdventOfCode/IntCode/Modules/Input/AmplifierInputModule.cs
@@ -4,7 +4,7 @@ namespace AdventOfCode.IntCode.Modules.Input
 {
     public class AmplifierInputModule : ListInputModule
     {
-        public AmplifierInputModule(int phaseMode, int input) : base(new List<int> {phaseMode, input})
+        public AmplifierInputModule(long phaseMode, long input) : base(new List<long> {phaseMode, input})
         {
         }
     }
