+++ AdventOfCode/IntCode/Modules/Output/ActionOutputModule.cs
@@ -5,14 +5,14 @@ namespace AdventOfCode.IntCode.Modules.Output
 {
     public class ActionOutputModule : IOutputModule
     {
-        private readonly Action<int> _callback;
+        private readonly Action<long> _callback;
 
-        public ActionOutputModule(Action<int> callback)
+        public ActionOutputModule(Action<long> callback)
         {
             _callback = callback;
         }
         
-        public void OutputCallback(int val)
+        public void OutputCallback(long val)
         {
             _callback.Invoke(val);
         }
